Remove alex checks
While the spirit of the checks is great the workflow is buggy to the
degree that it is more harmful then helpful. Nearly all warnings it
produces are false positives which leads to being desensitized to real
issues as they're just "clicked away" and human checks are not performed
as "the CI catches it".

Signed-off-by: Heinz N. Gies <heinz@licenser.net>